# OldProject/Common/LoginRequest.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace Common
{
    public class LoginRequest
    {
        public string Username { get; set; }
        public string password { get; set; }
    }
}