# aspnetcore-server-generated/Common/CalculationRequest.cs
namespace Common
{
    // Represents the input values for a calculation operation
    public class CalculationRequest
    {
        // The first number in the calculation
        public double Number1 { get; set; }

        // The second number in the calculation
        public double Number2 { get; set; }
    }
}